# rpi5_snapshot/openapi.yaml
openapi: 3.0.3
info:
  title: Game Stats API
  version: 1.0.0
  description: API kontrakty dla modułów 1v1 oraz multiplayer.
servers:
  - url: /api/v1
components:
  schemas:
    Uuid:
      type: string
      format: uuid
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
...
    ErrorDetailsItem:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorPayload:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, NOT_FOUND, CONFLICT, INTERNAL_ERROR]
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetailsItem'
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorPayload'
    MultiplayerGame:
      type: object
      required: [id, code, displayName, scoringType, minPlayers, maxPlayers, isActive, inOneVsOne, inMultiplayer, optionsCount, requiresOption, customFieldsCount]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
          example: "terraforming_mars"
        displayName:
          type: string
          example: "Terraformacja Marsa"
        scoringType:
          type: string
          enum: [MANUAL_POINTS, TTR_CALCULATOR, TM_CALCULATOR, CUSTOM_CALCULATOR]
        minPlayers:
          type: integer
        maxPlayers:
          type: integer
        isActive:
          type: boolean
        inOneVsOne:
          type: boolean
        inMultiplayer:
          type: boolean
        optionsCount:
          type: integer
        requiresOption:
          type: boolean
        customFieldsCount:
          type: integer
    MultiplayerGameListItem:
      type: object
      required: [id, code, displayName, scoringType, minPlayers, maxPlayers, isActive, inOneVsOne, inMultiplayer, optionsCount, requiresOption, customFieldsCount]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
        displayName:
          type: string
        scoringType:
          type: string
          enum: [MANUAL_POINTS, TTR_CALCULATOR, TM_CALCULATOR, CUSTOM_CALCULATOR]
        minPlayers:
          type: integer
        maxPlayers:
          type: integer
        isActive:
          type: boolean
        inOneVsOne:
          type: boolean
        inMultiplayer:
          type: boolean
        optionsCount:
          type: integer
        requiresOption:
          type: boolean
        customFieldsCount:
          type: integer
    MultiplayerGameList:
      type: array
      items:
        $ref: '#/components/schemas/MultiplayerGameListItem'
    MultiplayerGameCreateInput:
      type: object
      required: [displayName]
      properties:
        code:
          type: string
        displayName:
          type: string
        scoringType:
          type: string
          enum: [MANUAL_POINTS, CUSTOM_CALCULATOR]
        minPlayers:
          type: integer
        maxPlayers:
          type: integer
        inOneVsOne:
          type: boolean
        inMultiplayer:
          type: boolean
        customCalculator:
          type: object
          properties:
            fields:
              type: array
              items:
                type: object
                required: [label, pointsPerUnit]
                properties:
                  code:
                    type: string
                  label:
                    type: string
                  description:
                    type: string
                  pointsPerUnit:
                    type: integer
    MultiplayerGamePatchInput:
      type: object
      minProperties: 1
      properties:
        displayName:
          type: string
          maxLength: 80
        minPlayers:
          type: integer
          minimum: 2
          maximum: 5
        maxPlayers:
          type: integer
          minimum: 2
          maximum: 5
        inOneVsOne:
          type: boolean
        inMultiplayer:
          type: boolean
    MultiplayerGameDeleteResponse:
      type: object
      required: [code, deletedMatches]
      properties:
        code:
          type: string
        deletedMatches:
          type: integer
    MultiplayerCustomCalculatorField:
      type: object
      required: [id, gameId, code, label, pointsPerUnit, sortOrder, isActive]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        gameId:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
        label:
          type: string
        description:
          type: string
          nullable: true
        pointsPerUnit:
          type: integer
        sortOrder:
          type: integer
        isActive:
          type: boolean
...
    MultiplayerGameOption:
      type: object
      required: [id, gameId, code, displayName, sortOrder, isActive]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        gameId:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
        displayName:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /multiplayer/games:
    get:
      summary: List multiplayer games
      parameters:
        - in: query
          name: includeInactive
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiplayerGameList'
              examples:
                list:
                  value:
                    - id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                      code: "terraforming_mars"
                      displayName: "Terraformacja Marsa"
                      scoringType: "TM_CALCULATOR"
                      minPlayers: 2
                      maxPlayers: 5
                      isActive: true
                      inOneVsOne: false
                      inMultiplayer: true
                      optionsCount: 6
                      requiresOption: true
                      customFieldsCount: 0
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create multiplayer game (manual or custom calculator)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiplayerGameCreateInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiplayerGame'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /multiplayer/games/{code}:
    get:
      summary: Get multiplayer game by code
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: includeInactive
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiplayerGame'
              examples:
                game:
                  value:
                    id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    code: "ticket_to_ride"
                    displayName: "Pociągi"
                    scoringType: "TTR_CALCULATOR"
                    minPlayers: 2
                    maxPlayers: 5
                    isActive: true
                    inOneVsOne: false
                    inMultiplayer: true
                    optionsCount: 10
                    requiresOption: true
                    customFieldsCount: 0
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      summary: Update multiplayer game settings
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiplayerGamePatchInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiplayerGame'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete manual multiplayer game and clear related stats
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiplayerGameDeleteResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /multiplayer/games/{code}/options:
    get:
      summary: List active options for multiplayer game
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MultiplayerGameOption'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /multiplayer/games/{code}/calculator-fields:
    get:
      summary: List custom calculator fields for multiplayer game
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: includeInactive
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MultiplayerCustomCalculatorField'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
...